academic_research_extraction/requirements.txt
# Core dependencies
httpx>=0.24.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0
orjson>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
duckdb>=0.8.0

# Data processing
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
scispacy>=0.5.3
nltk>=3.8.0
keybert>=0.7.0
bertopic>=0.15.0
umap-learn>=0.5.0
hdbscan>=0.7.0

# Text processing
langdetect>=1.0.9
unidecode>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
simhash>=2.1.0
datasketch>=1.6.0

# Graph processing
networkx>=3.1.0
pyvis>=0.3.0
neo4j>=5.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0
altair>=5.1.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
loguru>=0.7.0

# PDF processing
pymupdf>=1.23.0
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
